Move PeakFitEngine refs #219
